Add files via upload
The graph depicts how well the model predicts each class of Alzheimer’s progression. The color of each bar indicates the level of accuracy with which the model classified that category of image. The lighter blue indicates a higher level of accuracy and the darker blue indicates a lower level of accuracy.